unit-translation-component/requirements.txt
coverage==5.1
pytest==5.4.2
rdflib==5.0.0
requests==2.22.0
SPARQLWrapper==1.8.5
redis==3.5.3
pyspellchecker==0.5.4
progress==1.5
